Move SettlementEncoder into own file
While working with the settlement file I felt it got very big so I split
out the encoder into it's own file.
